requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
credentials==1.1
future==0.18.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
httplib2==0.22.0
idna==3.4
lxml==4.9.2
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
peewee==3.16.0
pefile==2021.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyMySQL==1.0.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
socks==0
soupsieve==2.4.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
